Change Union{Missing, _} to concrete types
Elimination of the Missing specifier should speed up addition of arrays
in NetRevenue as there's no need to check each value. This makes columns
into 'concrete' types.

Further, columns of Float32 were changed to be Float64, since that is
the system default-this should further lead to less costly type
conversion.